Extracted operations from operation.js
