+++ src/operations/index.js
@@ -1,0 +1,11 @@
+export {accountMerge} from './account_merge';
+export {allowTrust} from './allow_trust';
+export {bumpSequence} from './bump_sequence';
+export {changeTrust} from './change_trust';
+export {createAccount} from './create_account';
+export {createPassiveOffer} from './create_passive_offer';
+export {manageData} from './manage_data';
+export {manageOffer} from './manage_offer';
+export {pathPayment} from './path_payment';
+export {payment} from './payment';
+export {setOptions} from './set_options';
